// frontend/src/shared/api/generated/blog/models/upsertPostRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Niklabs.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UpsertPostRequestCoverImageUrl } from './upsertPostRequestCoverImageUrl';

export interface UpsertPostRequest {
  title: string;
  excerpt: string;
  contentMarkdown: string;
  coverImageUrl: UpsertPostRequestCoverImageUrl;
  isPublished: boolean;
}